worker/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.1
requests==2.31.0
schedule==1.2.0 